requirements.txt
aiofiles==23.2.1
aiohttp==3.8.5
aiosignal==1.3.1
altair==5.1.0
annotated-types==0.5.0
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
contourpy==1.1.0
cycler==0.11.0
dataclasses-json==0.5.14
exceptiongroup==1.1.3
faiss-cpu==1.7.4
fastapi==0.103.0
ffmpy==0.3.1
filelock==3.12.3
fonttools==4.42.1
frozenlist==1.4.0
fsspec==2023.6.0
gradio==3.41.2
gradio_client==0.5.0
greenlet==2.0.2
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
huggingface-hub==0.16.4
idna==3.4
importlib-resources==6.0.1
iniconfig==2.0.0
Jinja2==3.1.2
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
kiwisolver==1.4.5
langchain==0.0.274
langsmith==0.0.27
loguru==0.7.0
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib==3.7.2
multidict==6.0.4
mypy-extensions==1.0.0
mysql-connector==2.2.9
numexpr==2.8.5
numpy==1.25.2
openai==0.27.9
orjson==3.9.5
packaging==23.1
pandas==2.0.3
Pillow==10.0.0
pluggy==1.2.0
psycopg==3.1.10
psycopg-binary==3.1.10
psycopg-pool==3.1.7
pydantic==2.3.0
pydantic_core==2.6.3
pydub==0.25.1
pyparsing==3.0.9
pytest==7.4.0
pytest-asyncio==0.21.1
python-dateutil==2.8.2
python-multipart==0.0.6
pytz==2023.3
PyYAML==6.0.1
referencing==0.30.2
regex==2023.8.8
requests==2.31.0
rpds-py==0.10.0
semantic-version==2.10.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.20
starlette==0.27.0
tenacity==8.2.3
tiktoken==0.4.0
tomli==2.0.1
toolz==0.12.0
tqdm==4.66.1
typing-inspect==0.9.0
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
uvicorn==0.23.2
websockets==11.0.3
yarl==1.9.2
zipp==3.16.2
